resolve github comments
