+++ app/v3/service/internal/keeper/state_change.go
@@ -16,7 +16,6 @@ func (k Keeper) CompleteBatch(ctx sdk.Context, requestContext types.RequestConte
 
 	if len(requestContext.ModuleName) != 0 {
 		tags = tags.AppendTags(k.Callback(ctx, requestContextID))
-		k.Callback(ctx, requestContextID)
 	}
 
 	batchState := types.BatchState{
